Update frond side switching to correct the data encoding.
The fronds in both LF and RF store the same min-max vertex data, but store it in dicts with different keys. This causes other portions of the code to break, because they expect the side-specific encoding. This update switches the encoding of the dict when switching it from one side to another.