Add a unit test for really large ring graph planarity
We want to have a planarity test for really large ring graphs. However, we know that our current implementation will produce a recursion depth error on large graphs. Thus, we have a test that attempts to run, but doesn't error out if it encounters the recursion depth exceeded runtime error.